% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{completa_scores}
\alias{completa_scores}
\title{Merge OC or w-nominate estimates with biographic information on deputies}
\usage{
completa_scores(rc)
}
\arguments{
\item{rc}{A OC or W-nominated object with estimates}
}
\description{
Merge OC or w-nominate estimates with biographic information on deputies
}
