% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{pega_autor}
\alias{pega_autor}
\title{Get bills' authors}
\usage{
pega_autor(anos, filtra_deputados = FALSE, proponente_principal = TRUE)
}
\arguments{
\item{anos}{A vector of years}

\item{filtra_deputados}{If \code{TRUE}, returns only deputies}

\item{proponente_principal}{If \code{TRUE}, returns only main authors}
}
\description{
Get bills' authors
}
