% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech.R
\name{pega_discursos}
\alias{pega_discursos}
\title{Retrieve legislators' speeches}
\usage{
pega_discursos(id_deputado, legis)
}
\arguments{
\item{id_deputado}{A deputy's ID}

\item{legis}{A chamber identification}
}
\description{
Retrieve legislators' speeches
}
