% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes.R
\name{pega_orientacao}
\alias{pega_orientacao}
\title{Get parties' whip positions in rollcall votes}
\usage{
pega_orientacao(ano)
}
\arguments{
\item{ano}{Year.}
}
\description{
Get parties' whip positions in rollcall votes
}
