% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputies.R
\name{pega_parlamentares}
\alias{pega_parlamentares}
\title{Get biographic information on all deputies since the foundation of the Chamber}
\usage{
pega_parlamentares()
}
\description{
Get biographic information on all deputies since the foundation of the Chamber
}
