% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{pega_proposicoes}
\alias{pega_proposicoes}
\title{Get bills}
\usage{
pega_proposicoes(anos, filtra_tipo = TRUE, temas = FALSE, autores = FALSE)
}
\arguments{
\item{anos}{A vector of years}

\item{filtra_tipo}{If \code{TRUE}, returns only PL, PLP, MPV and PEC}

\item{temas}{If \code{TRUE}, returns bills' topics as variables}

\item{autores}{If \code{TRUE}, returns authors' informations as variables}
}
\description{
Get bills
}
