% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{pega_temas}
\alias{pega_temas}
\title{Get bills' topics}
\usage{
pega_temas(anos, wide = TRUE)
}
\arguments{
\item{anos}{A vector of years.}

\item{wide}{If \code{TRUE}, returns a wide table}
}
\description{
Get bills' topics
}
