% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{pega_tramitacoes}
\alias{pega_tramitacoes}
\title{Pega a tramitacao de mais de uma proposicao}
\usage{
pega_tramitacoes(ids)
}
\arguments{
\item{ids}{ID da proposicao}
}
\description{
Pega a tramitacao de mais de uma proposicao
}
