% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes.R
\name{pega_votacoes}
\alias{pega_votacoes}
\title{Get a list of all rollcalls in a given year}
\usage{
pega_votacoes(ano, plenario = TRUE)
}
\arguments{
\item{ano}{Year}

\item{plenario}{If \code{TRUE}, returns only rollcalls in the plenary}
}
\description{
Get a list of all rollcalls in a given year
}
