% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes.R
\name{pega_votos}
\alias{pega_votos}
\title{Get deputies' rollcall votes in a given year}
\usage{
pega_votos(ano, plenario = TRUE, orientacao = TRUE)
}
\arguments{
\item{ano}{Year.}

\item{plenario}{If \code{TRUE}, returns only rollcalls in the plenary}

\item{orientacao}{IF \code{TRUE}, adds parties' whip orientations}
}
\description{
Get deputies' rollcall votes in a given year
}
