% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rice_camara}
\alias{rice_camara}
\title{Calculate Parties' Rice Index}
\usage{
rice_camara(ano, unidade = "bimonth")
}
\arguments{
\item{ano}{Year.}

\item{unidade}{Unit to aggregate data.}
}
\description{
Calculate Parties' Rice Index
}
