% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tema_olb}
\alias{tema_olb}
\title{OLB's ggplot2}
\usage{
tema_olb(posicao = "none", base_size = 14, base_family = "worksans")
}
\arguments{
\item{posicao}{Legend position}

\item{base_size}{Text size}

\item{base_family}{Text font (uses showtext)}
}
\description{
OLB's ggplot2
}
