% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{votos_para_rollcall}
\alias{votos_para_rollcall}
\title{Transform vote data into rollcall objects}
\usage{
votos_para_rollcall(votos, obstrucao_nao = TRUE, ideal = TRUE)
}
\arguments{
\item{votos}{An object returned from \code{\link{pega_votos}}}

\item{obstrucao_nao}{If \code{TRUE}, codes obstructions as nay votes}

\item{ideal}{If \code{TRUE}, returns and object suited to analysis using Bayesian IRT}
}
\description{
Transform vote data into rollcall objects
}
\note{
The function uses the deputy-party-uf unity
}
